requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.12.1
urllib3==1.26.13
